Rearranging tech macros.
